sensor_node/requirements.txt
Adafruit-Blinka==8.55.0
adafruit-circuitpython-ads1x15==2.4.2
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-dht==4.0.7
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
certifi==2025.1.31
charset-normalizer==3.4.1
colorzero==2.0
gpiozero==2.0.1
idna==3.10
paho-mqtt==2.1.0
psutil==7.0.0
pyftdi==0.56.0
pyserial==3.5
pyusb==1.3.1
requests==2.32.3
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
sysv-ipc==1.1.0
typing_extensions==4.12.2
urllib3==2.3.0
